Moved about us page and images
