+++ src/components/organisms/index.ts
@@ -14,4 +14,3 @@ export * from "./InfoPage";
 export * from "./TrophyBoxesContainer";
 export * from "./SettingsPage";
 export * from "./AccessibilitySettings";
-export * from "./AboutUs";
